New Creator class which now can create an airport from given JSON-file

+++ src/InAirQueue.scala
@@ -1,6 +1,6 @@
 
 
-class InAirQueue(val altitude: Int, val c: Int, val idN: Int) extends Queue(c, idN) {
+class InAirQueue(val altitude: Double, val c: Int, val idN: Int) extends Queue(c, idN) {
 
   def descendPlane(airplane: Airplane, runwayNo: Int): Unit = {
     this.getPlane(airplane).descend(runwayNo)
